# Related Issues
#512, #513 

# Notes
With this PR, full 360° CylinderSegment case is caught internally and computed with the cylinder core function with the cut-out method. This improves computation time for this special case by almost an order of magnitude.

# Tasks
- [x] catch special 360° case
- [x] update tests
- [ ] update docs (https://magpylib.readthedocs.io/en/latest/examples/examples_01_complex_forms/#union-operation) @OrtnerMichael ?

# Example
```python
from magpylib.magnet import Cylinder, CylinderSegment

# ring from CylinderSegment
ring0 = CylinderSegment(magnetization=(0,0,100), dimension=(2,3,1,0,360))
ring1 = CylinderSegment(magnetization=(0,0,100), dimension=(2,3,1,0,359.999))

%timeit ring0.getB((1,2,3))
%timeit ring1.getB((1,2,3))
```
```log
879 µs ± 10.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
7.13 ms ± 83.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```